function/requirements.txt
azure-common==1.1.25
azure-servicebus==7.0.0
cryptography==2.9.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary
python-http-client==3.2.7
requests==2.23.0
requests-oauthlib==1.3.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
